Add files via upload
This code serves as a supplementary component to the steps analysed and detailed in the master's thesis report. For a more comprehensive explanation and in-depth discussion it, please refer to the corresponding sections in the report.

Thesis Abstract:
This thesis investigates the application and evaluation of dimensionality reduction techniques, with a particular emphasis on UMAP, to address the challenges of analysing and evaluating dimensionality reduction visualizations. A comparative study of different methods was conducted for the recognition of handwritten digits and facial-based images to assess each technique's ability to preserve critical structures and patterns.

The UMAP hyperparameter analysis was conducted to evaluate its impact on both, inter and intra-cluster relationships, examining the influence of $n\_neighbours$ and $min\_dist$ on projection variability and cluster compactness. It involved computing qualitative and quantitative comparisons, including evaluations of cluster separation, compactness, stability, and density.

Showing the difficulties of assessing how well a dimensionality reduction technique is working, and how sensitive the technique is to certain hyperparameters. Furthermore, it highlighted the interplay between local and global structures and the trade-offs between prioritizing local relationships versus capturing broader global structures. Providing insights into the parametrization of the algorithms for different data characteristics.

Moreover, this research studied whether it is beneficial to create a pipeline combining multiple dimensionality reduction techniques with different characteristics, or if it is better to use a single technique.

In the domain of image-based facial emotion recognition, this research explored the impact of preprocessing techniques, including LLE, PCA and Gabor filters, on the quality of the lower dimension space. These methods showed their ability to enhance clustering metrics and refine data representation, particularly when combined with supervised learning. Additionally, an exploratory analysis investigated UMAP's capacity to identify image sources in merged datasets, illustrating its potential for domain identification and transfer learning applications.

The objectives of the study included enhancing visualization techniques through hierarchical architectures and analysing the role of global parameters. The findings serve as a foundation for understanding UMAP's behaviour and capabilities across various applications, providing insights and orientations for researchers to optimize the algorithm for improved stability, interpretability, and effectiveness in diverse high-dimensional data scenarios.